Rewrite source address with a proxy address
When natting a destination address it is also necessary to rewrite the
source address with a proxy address.

Example:

Masqueraded host: 10.10.10.7/192.168.0.2.

Init:
    - 10.10.10.2 -> 10.10.10.7.

Rewrite dest address:
    - 10.10.10.2 -> 192.168.0.2.

In order to reach its final destination, the packet should pretend to be
in the same network:
    - 192.168.0.1 -> 192.168.0.2.

The packet reaches its destination. It should reply back to its sender.
    - 192.168.0.2 -> 192.168.0.1.

But 192.168.0.1 is actually 10.10.10.2, so destination address is
rewritten.
    - 192.168.0.2 -> 10.10.10.2.

In order to reach its final destination, the packet should pretend to be
in the same network:
    - 10.10.10.1 -> 10.10.10.1.