Script to add 802.1q VLAN tagging to pcap files
This script uses the Scapy Python module to read packets from a pcap input
file, add 802.1q fields to their Ethernet frames, and write the modified
packets to another pcap file. The VLAN identifier can be optionally passed
as the last script parameter; if not specified "42" is used.

As a bonus, a Makefile which can be used to add 802.1q to the pcap files
under "tests/apps/lwaftr/data/" is included. The VLAN_ID environment
variable can be used to specify which VLAN identifier to use.