Add if-feature support to yang schema
This adds functioning support for the 'if-feature' statement. This
allows for programs to specify the features they support and have
the sections which pertain to unsupported features disabled.

The change is achived by changing the load_schema* commands to return
a getter, this getter takes a lua table as a list. The table should
list all the features it supports and the getter will return the
schema with those configured. If the getter is called without any
arguments, it's the same as not listing any supported features.