Fix VLAN tag setting after a NIC app is restarted
It's not enough to scan and find the first empty spot or
a matching VLAN in one pass, since depending on the order
that apps are shut down an earlier space may free up. Duplicate
VLAN entries in PFVLVF do not work, so we have to make sure
VLANs are unique.