Convert lwAFTR app to pipeline-of-queues style
The idea here is that we should write code that plays to the strengths
of LuaJIT, and LuaJIT's strength is tight and ideally branchless loops.
In this style, we move packets from queue to queue.  Because the packets
are alive in only a straight-line code path, LuaJIT's allocation sinking
optimization does very well for this style of code.

Unfortunately this patch is currently a regression, relative to the
upstream lwAFTR branch, so this is a work-in-progress patch set.