Disable blacklisting of problematic initial trace instructions
We have seen some cases where an important trace gets blacklisted just
because it hits an inner loop that hasn't compiled yet.  Now that
RaptorJIT has switched to a more hotcount-per-second model rather than a
cumulative-hotcount model, the eternal aspect of the blacklist no longer
matches the optimization paradigm we are going for.

Therefore add a "blacklist" JIT option off by default.  Instructions
will only ever be blacklisted when "blacklist" is on.